Add GitHub tool configuration component
